requirements.txt
aiohttp>=3.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
web3>=6.11.3
eth-account>=0.9.0
eth-typing>=3.0.0
typing-extensions>=4.5.0
solana>=0.35.1
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
backoff>=2.2.0
